tests: fix build on musl-based systems
Musl libc does not provide the backtrace functions in its implementation.
These functions are typically provided by glibc's <execinfo.h>.

As a fallback for musl-based systems, support linking against the external
libunwind library and using dladdr() to obtain similar stack trace info.

Note that the usage of dladdr() is similar to that of backtrace_symbols in
glibc, and it can be used with glibc as well.

Signed-off-by: Alexey Kodanev <aleksei.kodanev@bell-sw.com>